modules/crypto/crypto-cipher-aes/pyproject.toml
[tool.poetry]
name = "toku-crypto-cipher-aes"
version = "0.1.0"
description = "Project that provides the AES implemention for operating with cipher"
license = "Internal"
authors = [
    "Toku Dev <dev@trytoku.com>"
]
maintainers = [
    "Toku Dev <dev@trytoku.com>"
]
readme = "README.md"
homepage = "https://www.jira.com"
repository = "https://www.github.com"
documentation = "https://www.notion.com"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries"
]
packages = [
    {include = "toku"}
]

[tool.poetry.dependencies]
python = "^3.11"
overrides = "^7.4.0"
pycryptodome = "~3.19.0"
toku-crypto-cipher-api = "~0.1.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.1"
flake8 = "^6.1.0"
mypy = "^1.5.1"
black = "^23.9.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.3.1"
flexmock = "^0.11.3"
faker = "^19.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
